refactor: rename DomainPreview.tsx -> PreviewCard.tsx
